Rewrote nodes.py, removed corresponding tests, added license
nodes.py was unruly and incorrect for what was needed from a node,
it also needed new types of nodes. Node was changed from an old style
object to a new style object, keeping things "pythonic".

test_nodes.py will still need to be populated with adequate tests, which
do not exist at this time, but old tests were removed to prevent confusion
in the future. They were not commented out because they no longer had
relevance to what nodes.py is.

License text is provided, as encouraged by the FSF.